+++ levels/commit_in_future.rb
@@ -1,7 +1,12 @@
-require 'time'
+# require 'time'
+require 'date'
 
 difficulty 2
-description "Commit your changes with the future date (e.g. tomorrow)."
+# description "Commit your changes with the future date (e.g. tomorrow)."
+text = <<~TEXT
+  Speichern Sie Ihre Änderungen mit einem zukünftigen Datum (z.B. morgen).
+TEXT
+description text
 
 setup do
   repo.init
@@ -11,9 +16,12 @@ setup do
 end
 
 solution do
-  repo.commits.length == 1 && repo.commits.first.authored_date > Time.now
+  repo.commits(repo.head.name).length == 1 &&
+  repo.commits(repo.head.name).first.authored_date > Time.now
 end
 
 hint do
-  puts "Build a time format, and commit your code using --date parameter for \"future\"."
+  # puts "Build a time format, and commit your code using --date parameter for \"future\"."
+  datum = Date.today + 1
+  puts "Erstellen Sie einen Commit z.B. mit dem Parameter '--date #{datum}'."
 end