+++ levels/stash.rb
@@ -1,5 +1,10 @@
 difficulty 2
-description "You've made some changes and want to work on them later. You should save them, but don't commit them."
+# description "You've made some changes and want to work on them later. You should save them, but don't commit them."
+text = <<~TEXT
+  Sie haben einige Änderungen vorgenommen und wollen sie später weiter
+  bearbeiten. Sie sollten sie speichern, aber nicht im Repository eintragen.
+TEXT
+description text
 
 setup do
   init_from_level
@@ -12,5 +17,6 @@ solution do
 end
 
 hint do
-  puts "It's like stashing. Try finding an appropriate git command."
+  # puts "It's like stashing. Try finding an appropriate git command."
+  puts "Es ist wie ein Versteck. \nVersuchen Sie, einen geeigneten Git-Befehl zu finden."
 end